Constructor for DiagramElement is added.
Diagram elements that corresponds to empty space (not filled with space
characters) initialized with 0 instead of ' '. This fact broke
DiagramElement.isCharacter() method. Although it was possible to fix
this method only, the more consisten solution is to make default
constructor, which guaranties that there is no need to add weird checks
`c == ' ' and c == 0` all around the code.